Move parser fixtures to files
