+++ src/utils/tests/fixtures/class.js
@@ -1,0 +1,13 @@
+class Test {
+  constructor() {
+    this.foo = "foo";
+  }
+
+  bar(a) {
+    console.log("bar", a);
+  }
+}
+
+class Test2 {}
+
+let expressiveClass = class {};
