+++ src/utils/tests/fixtures/expression.js
@@ -1,0 +1,6 @@
+function expr() {
+  const obj = { a: { b: 2 } };
+  const obj2 = { c: { b: 3 } };
+  const foo = obj2.c.b;
+  return obj.a.b;
+}
