+++ src/utils/tests/fixtures/func.js
@@ -1,0 +1,9 @@
+function square(n) {
+  return n * n;
+}
+
+child = function() {};
+
+(function() {
+  2;
+})();
