+++ src/utils/tests/fixtures/math.js
@@ -1,0 +1,11 @@
+function math(n) {
+  function square(n) {
+    n * n;
+  }
+  const two = square(2);
+  const four = squaare(4);
+  return two * four;
+}
+
+var child = function() {};
+child2 = function() {};
