+++ src/utils/tests/fixtures/proto.js
@@ -1,0 +1,14 @@
+const foo = function() {};
+
+const bar = () => {};
+
+const TodoView = Backbone.View.extend({
+  tagName: "li",
+  initialize: function() {},
+  doThing(b) {
+    console.log("hi", b);
+  },
+  render: function() {
+    return this;
+  }
+});
