+++ src/utils/tests/fixtures/thisExpression.js
@@ -1,0 +1,11 @@
+class Test {
+  constructor() {
+    this.foo = {
+      a: "foobar"
+    };
+  }
+
+  bar() {
+    console.log(this.foo.a);
+  }
+}
