+++ src/utils/tests/fixtures/var.js
@@ -1,0 +1,4 @@
+var foo = 1;
+let bar = 2;
+const baz = 3;
+const a = 4, b = 5;
